fix: paginators no read stale state
This resolves issues where if you had code like:

```
onChange: (page, pageSize) => {
    paginator.setPage(page);
    if (pageSize) {
        paginator.setPageSize(pageSize);
    }
},
```

then the `setPageSize` would revert the call to `setPage` resulting
in the page never changing. This was because `setPageSize` reads
the current state internally. The implementation is such that the
easiest thing to do to maintain the current API is to just write
intermediate values to `this.currentState` directly. This should be
safe as `replaceStateControllers` should be called once the
`setState` is committed which will then set `this.currentState` to
the 'official' final value.