fix: handle empty many related fields
Previously we assumed that if we had an empty many related field ([])
then we just wouldn't record the nested fields because they can't be
inferred from an empty list. This caused problems when cloning a record
(or matching against '*') in that it would fail on nested records even
when it should really match (it should just return an empty array).

Instead we now default to the opposite - if no records then treat it
as all fields.